+++ components/commune/bal-anomalies.js
@@ -4,7 +4,6 @@ import theme from '@/styles/theme'
 
 import Anomaly from './anomaly'
 import Notification from '@/components/notification'
-import SectionText from '@/components/section-text'
 
 function BalAnomalies({errors}) {
   return (
@@ -16,17 +15,9 @@ function BalAnomalies({errors}) {
           </Notification>
         </div>
       ) : (
-        <>
-          <div className='anomalies-container'>
-            {errors.map(error => <Anomaly key={error.code} error={error.code} />)}
-          </div>
-          <div className='download-bal-container'>
-
-            <SectionText>
-              La Base Adresse Locale comporte des <b>anomalies</b>. Pour obtenir plus de détails concernant les alertes, vous avez la possibilité de <b>télécharger le fichier de la BAL</b> au format CSV et utiliser l’outils <a href='/bases-locales/validateur'>Validateur Base Adresse Locale</a>.
-            </SectionText>
-          </div>
-        </>
+        <div className='anomalies-container'>
+          {errors.map(error => <Anomaly key={error.code} error={error.code} />)}
+        </div>
       )}
 
       <style jsx>{`
@@ -55,17 +46,6 @@ function BalAnomalies({errors}) {
           align-items: center;
           gap: 1em;
         }
-
-        .download-bal-container {
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-        }
-
-       h4 {
-         margin-top: 2em;
-         text-align: center;
-       }
       `}</style>
     </div>
   )
