+++ lib/api-ban.js
@@ -54,7 +54,3 @@ export function getLieuxDitsCSVLegacy(codeCommune) {
   return `${API_BAN_URL}/ban/communes/${codeCommune}/download/csv-legacy/lieux-dits`
 }
 
-export function getCommuneCSVBAL(codeCommune) {
-  return `${API_BAN_URL}/ban/communes/${codeCommune}/download/csv-bal`
-}
-
